# Conflicts:
#	code/modules/admin/admin.dm

+++ code/modules/mob/living/carbon/human/species/station/ipc/ipc.dm
@@ -6,7 +6,7 @@
 	age_min = 1
 	age_max = 30
 	economic_modifier = 3
-	var/neuter_ipc = TRUE
+	default_genders = list(NEUTER)
 
 	blurb = "IPCs are, quite simply, \"Integrated Positronic Chassis.\" In this scenario, 'positronic' implies that the chassis possesses a positronic processing core (or positronic brain), meaning that an IPC must be positronic to be considered an IPC. The Baseline model is more of a category - the long of the short is that they represent all unbound synthetic units. Baseline models cover anything that is not an Industrial chassis or a Shell chassis. They can be custom made or assembly made. The most common feature of the Baseline model is a simple design, skeletal or semi-humanoid, and ordinary atmospheric diffusion cooling systems."
 
@@ -19,9 +19,9 @@
 	light_power = 0.5
 	meat_type = /obj/item/stack/material/steel
 	unarmed_types = list(
-		/datum/unarmed_attack/punch,
-		/datum/unarmed_attack/stomp,
-		/datum/unarmed_attack/kick)
+		/datum/unarmed_attack/punch/ipc,
+		/datum/unarmed_attack/stomp/ipc,
+		/datum/unarmed_attack/kick/ipc)
 	rarity_value = 2
 
 	inherent_eye_protection = FLASH_PROTECTION_MAJOR
@@ -38,6 +38,10 @@
 
 	brute_mod = 1.0
 	burn_mod = 1.2
+
+	grab_mod = 1.1 // Smooth, no real edges to grab onto
+	resist_mod = 2 // Robotic strength
+
 	show_ssd = "flashing a 'system offline' glyph on their monitor"
 
 	death_message = "gives one shrill beep before falling lifeless."
@@ -53,23 +57,24 @@
 	cold_level_2 = -1
 	cold_level_3 = -1
 
-	heat_level_1 = 500		// Gives them about 25 seconds in space before taking damage
-	heat_level_2 = 1000
-	heat_level_3 = 2000
+	heat_level_1 = 600		
+	heat_level_2 = 1200
+	heat_level_3 = 2400
 
 	body_temperature = null
 	passive_temp_gain = 10  // This should cause IPCs to stabilize at ~80 C in a 20 C environment.
 
 	inherent_verbs = list(
 		/mob/living/carbon/human/proc/self_diagnostics,
-		/mob/living/carbon/human/proc/change_monitor
+		/mob/living/carbon/human/proc/change_monitor,
+		/mob/living/carbon/human/proc/check_tag
 	)
 
 	flags = IS_IPC
 	appearance_flags = HAS_SKIN_COLOR | HAS_HAIR_COLOR
 	spawn_flags = CAN_JOIN | IS_WHITELISTED | NO_AGE_MINIMUM
 
-	blood_color = "#1F181F"
+	blood_color = COLOR_IPC_BLOOD
 	flesh_color = "#575757"
 	virus_immune = 1
 	reagent_tag = IS_MACHINE
@@ -98,7 +103,7 @@
 	)
 
 
-	heat_discomfort_level = 373.15
+	heat_discomfort_level = 500 //This will be 100 below the first heat level
 	heat_discomfort_strings = list(
 		"Your CPU temperature probes warn you that you are approaching critical heat levels!"
 		)
@@ -107,7 +112,7 @@
 
 	max_hydration_factor = -1
 
-	allowed_citizenships = list(CITIZENSHIP_NONE, CITIZENSHIP_BIESEL, CITIZENSHIP_FRONTIER, CITIZENSHIP_ERIDANI)
+	allowed_citizenships = list(CITIZENSHIP_NONE, CITIZENSHIP_BIESEL, CITIZENSHIP_COALITION, CITIZENSHIP_ERIDANI)
 	default_citizenship = CITIZENSHIP_NONE
 
 	// Special snowflake machine vars.
@@ -115,8 +120,6 @@
 	var/sprint_charge_factor = 0.65
 
 datum/species/machine/handle_post_spawn(var/mob/living/carbon/human/H)
-	if (neuter_ipc)
-		H.gender = NEUTER
 	. = ..()
 	check_tag(H, H.client)
 
@@ -144,28 +147,24 @@ datum/species/machine/handle_post_spawn(var/mob/living/carbon/human/H)
 	return sanitizeName(new_name, allow_numbers = 1)
 
 /datum/species/machine/proc/check_tag(var/mob/living/carbon/human/new_machine, var/client/player)
-	if (!new_machine || !player)
+	if(!new_machine || !player)
+		var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
+		if(tag)
+			tag.serial_number = uppertext(dd_limittext(md5(new_machine.real_name), 12))
+			tag.ownership_info = IPC_OWNERSHIP_COMPANY
+			tag.citizenship_info = CITIZENSHIP_BIESEL
 		return
 
-	if (establish_db_connection(dbcon))
+	var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
 
-		var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
-
-		var/status = TRUE
-		var/list/query_details = list("ckey" = player.ckey, "character_name" = player.prefs.real_name)
-		var/DBQuery/query = dbcon.NewQuery("SELECT tag_status FROM ss13_ipc_tracking WHERE player_ckey = :ckey: AND character_name = :character_name:")
-		query.Execute(query_details)
-
-		if (query.NextRow())
-			status = text2num(query.item[1])
-		else
-			var/DBQuery/log_query = dbcon.NewQuery("INSERT INTO ss13_ipc_tracking (player_ckey, character_name, tag_status) VALUES (:ckey:, :character_name:, 1)")
-			log_query.Execute(query_details)
-
-		if (!status)
-			new_machine.internal_organs_by_name -= BP_IPCTAG
-			new_machine.internal_organs -= tag
-			qdel(tag)
+	if(player.prefs.machine_tag_status)
+		tag.serial_number = player.prefs.machine_serial_number
+		tag.ownership_info = player.prefs.machine_ownership_status
+		tag.citizenship_info = new_machine.citizenship
+	else
+		new_machine.internal_organs_by_name -= BP_IPCTAG
+		new_machine.internal_organs -= tag
+		qdel(tag)
 
 /datum/species/machine/proc/update_tag(var/mob/living/carbon/human/target, var/client/player)
 	if (!target || !player)
@@ -296,8 +295,6 @@ datum/species/machine/handle_post_spawn(var/mob/living/carbon/human/H)
 /datum/species/machine/before_equip(var/mob/living/carbon/human/H)
 	. = ..()
 	check_tag(H, H.client)
-	if (neuter_ipc)
-		H.gender = NEUTER
 
 /datum/species/machine/handle_death_check(var/mob/living/carbon/human/H)
 	if(H.get_total_health() <= config.health_threshold_dead)
